feat: expose RustQC as a library crate
Adds a [lib] target alongside the existing [[bin]] so RustQC's analysis
modules can be consumed as a Rust library, not just via the CLI.

- New src/lib.rs publishes config, cpu, gtf, io, rna, summary as the
  public API and hosts the Strandedness enum at the crate root.
- main.rs is slimmed to bin-only modules (cli, ui, citations) and pulls
  the rest from rustqc::*.
- Strandedness moves out of cli.rs (which now imports it from the lib)
  so the 6 analysis modules using it no longer depend on the CLI module.
- format_count, format_pct, format_duration move from ui.rs to io.rs so
  library consumers can reach them; ui.rs uses them privately.

Closes #72
